requirements.txt
# Python dependencies for the market data pipeline

# =============================================================================
# WebSocket & HTTP
# =============================================================================
websockets>=12.0
aiohttp>=3.9.0
requests>=2.31.0

# =============================================================================
# Kafka
# =============================================================================
confluent-kafka>=2.3.0
fastavro>=1.9.0

# =============================================================================
# Spark (PySpark)
# =============================================================================
pyspark>=3.5.0

# =============================================================================
# Cassandra
# =============================================================================
cassandra-driver>=3.28.0

# =============================================================================
# Data Processing
# =============================================================================
pandas>=2.1.0
numpy>=1.26.0

# =============================================================================
# Configuration & Environment
# =============================================================================
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# =============================================================================
# Timezone Handling
# =============================================================================
pytz>=2023.3

# =============================================================================
# Logging & Monitoring
# =============================================================================
structlog>=23.2.0
prometheus-client>=0.19.0

# =============================================================================
# Testing
# =============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# =============================================================================
# Code Quality
# =============================================================================
black>=23.12.0
isort>=5.13.0
flake8>=6.1.0
mypy>=1.7.0

# =============================================================================
# Development
# =============================================================================
ipython>=8.18.0
jupyter>=1.0.0
